+++ task-shared-domain/src/main/java/ru/tinkoff/summer/taskshareddomain/ConnectionConstants.java
@@ -7,4 +7,6 @@ public class ConnectionConstants {
     public static final int RESULT_TOPIC_PARTITIONS = 3;
     public static final String EXECUTOR_GROUP_ID ="task-executor";
     public static final String SERVICE_GROUP_ID ="task-service";
+    //TODO: Профили тут, task-microservice, docker-compose. для dev - localhost, в остальном - kafka
+    public static final String BOOTSTRAP_SERVERS = "localhost:9092";
 }
