+++ task-shared-domain/src/main/java/ru/tinkoff/summer/taskshareddomain/TotalExecutionResult.java
@@ -13,7 +13,7 @@ import java.util.List;
 @ToString
 @Builder
 public class TotalExecutionResult {
-    long attemptId;
+    Long attemptId;
     ExecutionStatus status;
     double executionTimeNs;
     String actualResult;
